requirements.txt
Django==1.6
django-tastypie==0.11.0
ipython==1.1.0
pyreadline==2.0
python-dateutil==2.2
python-mimeparse==0.1.4
six==1.4.1
